+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/dto/ChangeDTO.java
@@ -19,11 +19,11 @@ import java.sql.Date;
 @Component
 @JsonComponent
 public class ChangeDTO {
-    private int changeId;
+    private Long changeId;
     private String changeStatus;
     private Date startDate;
     private Date endDate;
     private UniversityClass universityClass;
-    private int studentId;
-//    private int messageId;
+    private Long studentId;
+    //private int messageId;
 }
