+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/dto/MessageDTO.java
@@ -19,9 +19,9 @@ import java.sql.Time;
 @JsonComponent
 public class MessageDTO {
 
-    private int messageId;
+    private Long messageId;
     private Date messageDate;
     private Time messageTime;
     private String messageText;
-    private int changeId;
+    private Long changeId;
 }
